Homework sprint 11, changes after review
