+++ app/build.gradle.kts
@@ -51,6 +51,7 @@ dependencies {
     implementation("com.github.bumptech.glide:glide:4.14.2")
     annotationProcessor("com.github.bumptech.glide:compiler:4.14.2")
 
-    implementation("com.squareup.retrofit2:retrofit:2.9.0")
-    implementation ("com.squareup.retrofit2:converter-gson:2.9.0")
+    implementation("com.squareup.retrofit2:retrofit:${libs.versions.retrofit.get()}")
+    implementation ("com.squareup.retrofit2:converter-gson:${libs.versions.retrofit.get()}")
+
 }
