+++ app/src/main/java/com/practicum/playlistmaker/TrackSearchApi.kt
@@ -1,0 +1,11 @@
+package com.practicum.playlistmaker
+
+import retrofit2.Call
+import retrofit2.http.POST
+import retrofit2.http.Query
+
+interface TrackSearchApi {
+
+    @POST("/search?entity=song")
+    fun searchTrack(@Query("term") text: String): Call<TracksSearchResponse>
+}
